Don't bring symbols forward to non-existing phases.
This was observed in TabCompleterTests. The problem in general is that
we might see symbols in runs that have fewer phases than the phase at which the
symbol was created. Previously, we updated the symbol anyway. But this
is problematic since it means the symbol has a validity period that
does not correspond to a phase in the current run.

We now treat those symbols as stale without a way to recover instead.