Simplify Command API
Replace `argGetter`/`varargsGetter` with `parseArg`/`parseVararg`.
This new version returns an `Option[T]` instead of a `() => T`.
`None` can be used to state that this argument could not be parsed and
continue with the next `parseArg`.

The expectation is that `run` will first check if there was a parsing
failure and then call the main methods (just as we do with the lambdas).