Introduce `boundary/break` control abstraction.
The abstractions are intended to replace the `scala.util.control.Breaks` and
`scala.uitl.control.NonLocalReturns`. They are simpler, safer, and more performant,
since there is a new MiniPhase `DropBreaks` that rewrites local breaks to labeled
returns, i.e. jumps.

The abstractions are not experimental. This break from usual procedure is because
we need to roll them out fast. Non local returns were just deprecated in 3.2, and
we proposed `NonLocalReturns.{returning,throwReturn}` as an alternative. But these
APIs were a mistake and should be deprecated. So rolling out boundary/break now counts
as a bugfix.