Drop distinction between filterable and normal sync channels
Filterable sync channels don't require significantly more complexity than
normal channels. The only thing a normal channel could do that a filterable
channel could not, was to discard a reader and sender immediately if there
was no match. If there are no filters on the source, then the only way a
source could not match is by losing a race. But then the source will be
removed anyway by the `race` method. So the only improvement would be had
if there was a data race condition where the removal by race happens after
a second channel was triggered.