Rename CancellationGroup to CompletionGroup
Also, add a way to install completion handlers that are called when
a member of a completion group completes. This allows us to define
supervisors for coroutines or actors. Note that, unlike for futures
we sometimes do not want to wait for termination of a coroutine (or
actor) in a specific val. Instead, the coroutine could be terminated
when the enclosing completion group completes (i.e. gets cancelled).
But we still want to handle abnormal early termination of the coroutine
somehow. Completion handlers are the solution to this. A handler
can be installed to be valid in a scope. Then, if any cancellable objects
terminate, the handler is called.