Forbid defining inline traits inside inline traits
When inlining the body of an inline trait, trying to inline another inline
trait creates typing issues.

For example, with the following code:
```
inline trait A[T]:
  inline trait InnerA[U]:
      val x: (T, U) = ???

class B extends A[Int]:
  class InnerB extends InnerA[String]
```
when the Inlining phase tries to inline InnerA inside of InnerB, it
should try to change A#T to B#T but does not know that A exists in a
higher scope.

An idea to solve this would be to inline everything from A inside of B
before inlining InnerA inside of InnerB, but this adds quite a bit of
complexity.
Another one would be to keep track, when inlining a trait, of all the
type parameters like we do for term parameters. This could be done more
easily.

This will be left as future work for the time being.